configure zeitwerk
