+++ cloudflare-rails.gemspec
@@ -34,6 +34,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency "railties", ">= 6.1", "< 7.2.0"
   spec.add_dependency "activesupport", ">= 6.1", "< 7.2.0"
   spec.add_dependency "actionpack", ">= 6.1", "< 7.2.0"
+  spec.add_dependency "zeitwerk", ">= 2.5.0"
 
   # we need Module#prepend
   spec.required_ruby_version = '>= 2.0'
